requirements.txt
# Core AI/ML
torch
torchvision
transformers
openai

# Computer Vision
opencv-python-headless
pillow
easyocr
ultralytics

# Speech processing
speechrecognition
pyttsx
gtts

# Audio processing (for sound localization - v1.1+)
librosa>=0.10.0
scipy>=1.10.0

# APIs & Web
fastapi
uvicorn[standard]
requests

# Database
sqlalchemy
alembic

# Utilities
python-dotenv
pyyaml
numpy

# Optional (comment if not needed)
pyaudio # Might need manual installation
geopy # For navigation features
langchain